pyproject.toml
[tool.poetry]
name = "clang_tidy_checker"
version = "0.3.0"
description = "Tool to check C / C++ source codes using clang-tidy."
authors = ["Kenta Kabashima <kenta_program37@hotmail.co.jp>"]
license = "MIT"
readme = "README.md"
homepage = "https://gitlab.com/MusicScience37Projects/tools/clang-tidy-checker"
repository = "https://gitlab.com/MusicScience37Projects/tools/clang-tidy-checker.git"
documentation = "https://musicscience37projects.gitlab.io/tools/clang-tidy-checker/"
keywords = ["clang-tidy", "c++"]
# Classifiers: https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: C",
    "Programming Language :: C++",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
PyYAML = ">=6"
click = ">=8.1.6"
tqdm = ">=4.65.0"
msgpack = ">=1.0.7"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
Sphinx = "^7.2.6"
black = "^23.12.1"
doc8 = "^1.1.1"
pylint = "^3.0.3"
sphinx-autobuild = "^2021.3.14"
sphinx-orange-book-theme = "^0.2.0"
mypy = "^1.8.0"
flake8 = "^7.0.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
cmakelang = "^0.6.13"
pytest-approvaltests = "^0.2.4"
types-PyYAML = ">=6"
toml = "^0.10.2"
types-toml = "^0.10.8.7"
myst-parser = "^2.0.0"
pytest-asyncio = "^0.23.3"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
disable = "too-many-instance-attributes,line-too-long,fixme"
load-plugins = "pylint.extensions.docparams"

[tool.poetry.scripts]
clang-tidy-checker = "clang_tidy_checker.main:main"

[tool.isort]
profile = "black"
